requirements.txt
tushare>=1.4.0
akshare>=1.18,<2  # upstream Eastmoney backend is flaky; pin a major
baostock
google-genai>=1.0  # successor to deprecated google-generativeai
python-dotenv>=1.0.0
requests>=2.31.0
